+++ workalendar/tests/test_asia.py
@@ -63,6 +63,30 @@ class ChinaTest(GenericCalendarTest):
         self.assertNotIn(date(2020, 9, 27), holidays)  # National Day Shift
         self.assertNotIn(date(2020, 10, 10), holidays)  # National Day Shift
 
+    def test_year_2021(self):
+        holidays = self.cal.holidays_set(2021)
+        self.assertIn(date(2021, 1, 1), holidays)  # New Year
+        self.assertIn(date(2021, 2, 11), holidays)  # Spring Festival
+        self.assertIn(date(2021, 2, 17), holidays)  # Spring Festival
+        self.assertIn(date(2021, 4, 3), holidays)  # Ching Ming Festival
+        self.assertIn(date(2021, 4, 5), holidays)  # Ching Ming Festival
+        self.assertIn(date(2021, 5, 1), holidays)  # Labour Day Holiday
+        self.assertIn(date(2021, 5, 5), holidays)  # Labour Day Holiday
+        self.assertIn(date(2021, 6, 12), holidays)  # Dragon Boat Festival
+        self.assertIn(date(2021, 6, 14), holidays)  # Dragon Boat Festival
+        self.assertIn(date(2021, 9, 19), holidays)  # Mid-Autumn Festival
+        self.assertIn(date(2021, 9, 21), holidays)  # Mid-Autumn Festival
+        self.assertIn(date(2021, 10, 1), holidays)  # National Day
+        self.assertIn(date(2021, 10, 7), holidays)  # National Day
+
+        self.assertNotIn(date(2021, 2, 7), holidays)  # Spring Festival Shift
+        self.assertNotIn(date(2021, 2, 20), holidays)  # Spring Festival Shift
+        self.assertNotIn(date(2021, 4, 25), holidays)  # Labour Day Shift
+        self.assertNotIn(date(2021, 5, 8), holidays)  # Labour Day Shift
+        self.assertNotIn(date(2021, 9, 18), holidays)  # Mid-Autumn Shift
+        self.assertNotIn(date(2021, 9, 26), holidays)  # National Day Shift
+        self.assertNotIn(date(2021, 10, 9), holidays)  # National Day Shift
+
     def test_missing_holiday_year(self):
         save_2018 = china_holidays[2018]
         del china_holidays[2018]
@@ -75,7 +99,7 @@ class ChinaTest(GenericCalendarTest):
         with patch('warnings.warn') as patched:
             self.cal.get_calendar_holidays(year)
         patched.assert_called_with(
-            'Support years 2018-2020 currently, need update every year.'
+            'Support years 2018-2021 currently, need update every year.'
         )
 
     def test_is_working_day(self):